Remove Error Configuration component and related routes from the application
