+++ client/src/Routes.jsx
@@ -5,7 +5,6 @@ import { Landing } from "./pages/Landing/Landing";
 
 // Import playground pages
 import { PlaygroundLanding } from "./pages/Playground/PlaygroundLanding/PlaygroundLanding";
-import { ErrorConfiguration } from "./pages/Playground/ErrorConfiguration/ErrorConfiguration";
 
 import { ErrorPage } from "./pages/Error/ErrorPage";
 import { Settings } from "./pages/Settings/Settings";
@@ -43,11 +42,6 @@ export const routes = [
     layout: "playground",
     routes: [
       { title: "Playground", path: "/playground", element: <PlaygroundLanding /> },
-      {
-        title: "Error configuration",
-        path: "/playground/error-configuration",
-        element: <ErrorConfiguration />,
-      },
     ],
   },
   {
